Move PropertyGroup classes to new package
