+++ java/configuration/src/main/java/sleeper/configuration/properties/SystemDefinedInstanceProperty.java
@@ -17,6 +17,8 @@ package sleeper.configuration.properties;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
+import sleeper.configuration.properties.group.InstancePropertyGroup;
+
 import static sleeper.configuration.properties.SystemDefinedInstancePropertyImpl.named;
 
 /**
