+++ java/configuration/src/main/java/sleeper/configuration/properties/group/PropertyGroupImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.configuration.properties;
+package sleeper.configuration.properties.group;
 
 import java.util.ArrayList;
 import java.util.Collections;
