+++ java/configuration/src/test/java/sleeper/configuration/properties/InstancePropertyGroupingTest.java
@@ -18,6 +18,8 @@ package sleeper.configuration.properties;
 
 import org.junit.jupiter.api.Test;
 
+import sleeper.configuration.properties.group.PropertyGroup;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -27,9 +29,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.configuration.properties.InstanceProperty.getAllGroupedProperties;
 import static sleeper.configuration.properties.InstanceProperty.getAllProperties;
 import static sleeper.configuration.properties.InstanceProperty.sortProperties;
-import static sleeper.configuration.properties.InstancePropertyGroup.BULK_IMPORT;
-import static sleeper.configuration.properties.InstancePropertyGroup.COMMON;
-import static sleeper.configuration.properties.InstancePropertyGroup.INGEST;
+import static sleeper.configuration.properties.group.InstancePropertyGroup.BULK_IMPORT;
+import static sleeper.configuration.properties.group.InstancePropertyGroup.COMMON;
+import static sleeper.configuration.properties.group.InstancePropertyGroup.INGEST;
 
 public class InstancePropertyGroupingTest {
     @Test
